Put all the logic in StreamDecoder
I wasn't liking how ProtocolMesg and StreamDecoder would both need to
have business logic - as was, ProtocolMesg was responsible for reading
fields out of the message buffer, while ProtocolMesgFactory or
StreamDecoder needed logic to parse messages, check checksums, and
create the messages.

The prior design also gave every ProtocolMesg a
full copy of the raw message as recieved, and that wasted a lot of space
(mostly in either allocating Vectors that weren't required or oversizing
buffers to fit large strings that aren't usually needed.

Now the ProtocolMesg family of classes are simple containers.
StreamDecoder will parse the stream and generate the classes, and the
container provides access to human-friendly fields.