Added sequence number sorting & message retrieval
To make the polymorphism and message handoff work, we need to return a
pointer to the message instead of a raw message. Otherwise C++ doesn't
know to allocate extra space for each of the subclasses fields.

Passing around pointers also means we allocate memory on the heap once,
but never use copy constructors to move elements into arrays and out
through function returns.