That's all!
This adds a bunch more tests and a little more behavior. The tests
confirm that fields can be read from each of the message types, that
messages are retrieved in a sensible order based on sequence number,
that the parent message class also has correct fields.

I'll probably clean up a little bit with comments & such, and update the
readme. Then this is ready to ship.